Revamped configuration parsing.
The command line and configuration file parsing is now common to both
the synchronizer and the query tools. It is flexible enough to support
any other tool built on top of pycarddav, or to easily add future
options.

Also, the configuration file format was modified to introduce
sections, separating DAV configuration from SSL options or SQLite
database path.