Added support for .netrc.
DAV credentials are now read from the configuration file, from the
.netrc file and from user input in this order. If credentials are
found in .netrc, then those credentials are used only if either no
user was provided in configuration, or if the provided user matches
the one found in the credential.

Consequently, the DAV user item is no more mandatory in the
configuration file. However, if .netrc is not used, then the user name
should be provided because only the password is prompted by the
program. Log messages are explicit.