Updated the configuration handling.
The Configuration Parser now relies on a collection of Section
Parsers. The Configuration Parser reads the file sections one at a
time, using the matching section parser. The produced data sets are a
wrapped in a Namespace, and can be unique or belong to a collection.
The Configuration Parser returns all those data sets in a top-level
Namespace. The data read on the command line is also stored in this
top-level Namespace. Its content depends on the tool used.

pycardsyncer is broken at the moment, since it is not aware of the new
WebDAV accounts configuration.