version number generation added
Version numbers need to be changed at only one place (setup.py).

If pycarddav is installed from a git repository, a precise version
number is generated, if this isn't a release '.dev' is appended (also
the non-release version is installed from a tarball etc.).

On install, pycarddav/version.py is generated containing the version
number.