Add new modes to query: name, fname and allnames
Previously, when using pc_query, the full vcard was searched for
matches, which can be very time consuming, specially if vcards contains
images for the contact. A new command line option and configuration
variable is added (--where and [query] where) to control where to
search. Possible options are vcard (the default and the previous
behaviour), name and fname (searching in those DB fields respectively)
and allnames (searching in both).

Using any of the options but vcard yields a reduction in search time by
almost 1/3 (400ms vs 175ms) when compared with the vcard search.

Improves #85.